Migration to complete MEAN stack 9/18/2016
This commit comes as I am about to begin work to migrate this project to a full
MEAN stack. As of right now, Mongo, Express, and NodeJS are in use, but Angular
is not a part of the project.

This branch will serve as a placeholder branch for the project should I ever want
to return to work immediately on the version of the project which uses Skeleton.

There are no substantial differences between this version and a version saved as
a previous commit on the branch master.